Move everything out of packs so that we won't compile unnecessary stuff
