+++ app/javascript/BuiltInFormControls/RegistrationPolicyEditor.jsx
