+++ app/javascript/BuiltInFormControls/ScheduledValueEditor.jsx
