+++ app/javascript/FormPresenter/components/MultipleChoiceItem.jsx
