+++ app/javascript/FormPresenter/components/RegistrationPolicyItem.jsx
