+++ app/javascript/FormPresenter/components/RequiredIndicator.jsx
