+++ app/javascript/FormPresenter/components/StaticTextItem.jsx
