+++ app/javascript/FormPresenter/reducers/apiConfigurationReducer.js
