+++ app/javascript/FormPresenter/reducers/sectionTraversalReducer.js
