+++ app/javascript/HTTPUtils.js
