+++ app/javascript/LoadingIndicator.jsx
