+++ app/javascript/Models/RegistrationPolicy.js
